# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/v1/market/charts": {
            "get": {
                "description": "Get the charts data from an market and coin/token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Get charts data for a specific coin",
                "operationId": "get_charts_data",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 60,
                        "description": "Coin ID",
                        "name": "coin",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1574483028,
                        "description": "Start timestamp",
                        "name": "time_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "USD",
                        "description": "The currency to show charts",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockatlas.ChartData"
                        }
                    }
                }
            }
        },
        "/v1/market/info": {
            "get": {
                "description": "Get the charts coin info data from an market and coin/contract",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Get charts coin info data for a specific coin",
                "operationId": "get_charts_coin_info",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 60,
                        "description": "Coin ID",
                        "name": "coin",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1574483028,
                        "description": "Start timestamp",
                        "name": "time_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "USD",
                        "description": "The currency to show coin info in",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockatlas.ChartCoinInfo"
                        }
                    }
                }
            }
        },
        "/v1/market/ticker": {
            "get": {
                "description": "Get the ticker value from an market and coin/token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "Get ticker value for a specific market",
                "operationId": "get_ticker",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "coin id",
                        "name": "coin",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token id",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "USD",
                        "description": "the currency to show the quote",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockatlas.Ticker"
                        }
                    }
                }
            },
            "post": {
                "description": "Get the ticker values from many markets and coin/token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "Get ticker values for a specific markets",
                "operationId": "get_tickers",
                "parameters": [
                    {
                        "description": "Ticker",
                        "name": "tickers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TickerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockatlas.Tickers"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Coin": {
            "type": "object",
            "properties": {
                "coin": {
                    "type": "integer"
                },
                "token_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.TickerRequest": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Coin"
                    }
                },
                "currency": {
                    "type": "string"
                }
            }
        },
...
        "blockatlas.ChartCoinInfo": {
            "type": "object",
            "properties": {
                "circulating_supply": {
                    "type": "number"
                },
                "market_cap": {
                    "type": "number"
                },
                "total_supply": {
                    "type": "number"
                },
                "volume_24": {
                    "type": "number"
                }
            }
        },
        "blockatlas.ChartData": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "prices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockatlas.ChartPrice"
                    }
                }
            }
        },
        "blockatlas.ChartPrice": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "blockatlas.Ticker": {
            "type": "object",
            "properties": {
                "coin": {
                    "type": "integer"
                },
                "coin_name": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "last_update": {
                    "type": "string"
                },
                "price": {
                    "type": "object",
                    "$ref": "#/definitions/blockatlas.TickerPrice"
                },
                "token_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "blockatlas.TickerPrice": {
            "type": "object",
            "properties": {
                "change_24h": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "blockatlas.Tickers": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/blockatlas.Ticker"
            }
        },
...
    }
}